move appliance fixup fixture to fixtures.appliance
