+++ conftest.py
@@ -37,59 +37,6 @@ def pytest_addoption(parser):
     parser.getgroup('cfme', 'cfme: options related to cfme/miq appliances')
 
 
-def pytest_sessionstart(session):
-    try:
-        store.current_appliance.check_no_conflicting_providers()
-    except ApplianceException as e:
-        raise pytest.UsageError("Conflicting providers were found: {}".format(e))
-
-
-@pytest.fixture(scope="session", autouse=True)
-def set_session_timeout():
-    store.current_appliance.set_session_timeout(86400)
-
-
-@pytest.fixture(scope="session", autouse=True)
-def ensure_websocket_role_disabled():
-    # TODO: This is a temporary solution until we find something better.
-    roles = configuration.get_server_roles()
-    if 'websocket' in roles and roles['websocket']:
-        logger.info('Disabling the websocket role to ensure we get no intrusive popups')
-        roles['websocket'] = False
-        configuration.set_server_roles(**roles)
-
-
-@pytest.fixture(scope="session", autouse=True)
-def fix_merkyl_workaround():
-    """Workaround around merkyl not opening an iptables port for communication"""
-    ssh_client = store.current_appliance.ssh_client
-    if ssh_client.run_command('test -s /etc/init.d/merkyl').rc != 0:
-        logger.info('Rudely overwriting merkyl init.d on appliance;')
-        local_file = data_path.join("bundles").join("merkyl").join("merkyl")
-        remote_file = "/etc/init.d/merkyl"
-        ssh_client.put_file(local_file.strpath, remote_file)
-        ssh_client.run_command("service merkyl restart")
-        art_client.fire_hook('setup_merkyl', ip=appliance_ip_address)
-
-
-@pytest.fixture(scope="session", autouse=True)
-def fix_missing_hostname():
-    """Fix for hostname missing from the /etc/hosts file
-
-    Note: Affects RHOS-based appliances but can't hurt the others so
-          it's applied on all.
-    """
-    ssh_client = store.current_appliance.ssh_client
-    logger.info("Checking appliance's /etc/hosts for its own hostname")
-    if ssh_client.run_command('grep $(hostname) /etc/hosts').rc != 0:
-        logger.info("Adding it's hostname to its /etc/hosts")
-        # Append hostname to the first line (127.0.0.1)
-        ret = ssh_client.run_command('sed -i "1 s/$/ $(hostname)/" /etc/hosts')
-        if ret.rc == 0:
-            logger.info("Hostname added")
-        else:
-            logger.error("Failed to add hostname")
-
 
 @pytest.fixture(autouse=True, scope="function")
 def appliance_police():
